+++ RecDP/tests/test_llmutils.py
@@ -74,22 +74,14 @@ class Test_LLMUtils(unittest.TestCase):
         text_to_jsonl_MP(data_dir, out_dir, 2)
 
     def test_ppi_remove(self):
-        from dataclasses import dataclass
+        from pyspark.sql import SparkSession
 
-        @dataclass
-        class PiiDetectRedactOption:
-            path: str = "json"
-            data_files: str = "tests/data/llm_data/arxiv_sample_100.jsonl"
-            split: str = "train"
-            text_column: str = "text"
-            batch_size: int = 100
-            num_proc: int = 8
-            seed: int = 10
-            save_path: str = "./pii_remove"
-            save_format: str = "json"
+        spark = SparkSession.builder.master("local[2]").appName(
+            "PII Remove").getOrCreate()
 
-        args = PiiDetectRedactOption()
-        pii_remove(args)
+        input_dataset = spark.read.load(path="tests/data/llm_data/arxiv_sample_100.jsonl", format="json")
+        output_dataset = pii_remove(input_dataset)
+        output_dataset.write.save(path="./tmp", format="json", mode="overwrite")
 
     def test_filter_jsonl(self):
         data_dir = "tests/data/llm_data"
